requirements.txt
# Reddit API
praw>=7.7.1

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Slack
slack-sdk>=3.27.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
structlog>=24.1.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
